requirements.txt
aiohttp==3.7.4.post0
anyio==3.2.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
discord.py==1.7.3
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
multidict==5.1.0
numpy==1.20.2
Pillow==8.2.0
pycparser==2.20
pygal==2.4.0
pygame==2.0.1
PyNaCl==1.4.0
PySocks==1.7.1
pytube==10.9.0
requests==2.25.1
rfc3986==1.5.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.3
wikipedia==1.4.0
yarl==1.6.3
youtube-search-python==1.4.6
